feat: create and use validation per user

+++ src/modules/shopping/application/delete-list-by-id/delete-list-by-id-use-case.ts
@@ -2,6 +2,7 @@ import { InvalidUserIdError } from 'modules/users/domain/errors/invalid-user-id-
 import { PostgresShoppingListRespository } from 'modules/shopping/infrastructure/database/postgres-shopping-list-repository';
 import { IDeleteListByIdDTO } from './delete-list-by-id-dto';
 import { ListNotFound } from 'modules/shopping/domain/errors/list-not-found';
+import { NoPermission } from '../../../../shared/errors/no-permission';
 
 class DeleteListByIdUseCase {
   constructor(private shoppingListRepository: PostgresShoppingListRespository) {}
@@ -12,6 +13,10 @@ class DeleteListByIdUseCase {
     const listExist = await this.shoppingListRepository.getListById(data.id);
 
     if (listExist === null) throw new ListNotFound();
+
+    if (listExist.userId !== data.userId) {
+      throw new NoPermission();
+    }
 
     await this.shoppingListRepository.deleteListById(data);
   }
